Check for Small Cells

Small Cells Check checks a given table for small sells then adds a smallCells table to the MetaData of the table object

# S3 method for TableOne
CheckSmallCells(passedTable, smallSize = 6,
  print = FALSE, tableType = "TableOne")

Arguments

passedTable
The object outputted by the CreateTableOne function of the tableone package. The documentation is available here https://cran.r-project.org/web/packages/tableone/index.html.

smallSize
What value constitutes a small size cell. Default value is 6.

print
If TRUE prints the smallSize metadata in a human readable format

tableType
Specifies the type of the table that is passed to the function

Value

The passedTable object with a new object in the Metadata object called smallCells. smallCells is a dataframe with 4 columns stratifiedBy : the categeries the table was stratified by strataValues : the strata value where the small cell is present variableName and factors and the rows are all the categorical variables whose one or more factors have small cells.

Details

Checks the categorical table within the TableOne param (CatTable field) for small cells. A small cell is a category where the number of people in the category (n) is less than the value specified by the smallSize param. The freq field within each variable has the n values.

Examples

# Read in the data we will use to generate Table One

library(survival)
data(pbc)

# Create the Table One object

library("tableone")
# The list of variables which are categorical
catVars <- c("status", "trt", "ascites", "hepato",
             "spiders", "edema", "stage")

# create table 1 object
TableOne <- CreateTableOne(data = pbc,strata = c("trt","stage"), factorVars = catVars)

# by default smallSize is 6 print is set to true and tableType is TableOne
tmp <- CheckSmallCells(TableOne)

#> 8  variables with  160  levels checked.
#> 
#> 8  variables with  79  levels have cells < 6  counts.
#> 

# increasing the smallSize threshold to 10
tmp <- CheckSmallCells(TableOne, smallSize=10)

#> 8  variables with  160  levels checked.
#> 
#> 8  variables with  90  levels have cells < 10  counts.
#> 

# currently only TableOne is supported so tableType != TableOne will throw error
#tmp <- CheckSmallCells(TableOne, tableType="TableTwo")