Assign the earliest event, when there are multiple events

Finds the earliest date and returns the event and date for that event. Use when there are multiple events and dates. Each row represents an observation. The output of this function creates a variable that is used for survival models and predictive algorithms.

Input are lists of event dates for the main event (outcome), competing event (optional) and withdraw event (optional). The maximum follow-up time must be specified (followup_time).

The lists should be the same length but, if not, the return list is the same length as the longest input list.

Note: if making events using `make_event_dates`. The`make_event_dates` function randomly assigns event dates and therefore there will be withdraw events that 'occur' before 'main events'. This means that the main event rate will be lower than the original starting data.

assign_first_event(main_events, competing_events = NA,
  withdraw_events = NA, followup_time, main_label = "main event",
  competing_label = "competing event",
  withdraw_label = "withdraw from study", units = NA)

Arguments

main_events
The main outcome or event. A list of numbers (typically integers) that represent a unit of time (typically days).

competing_events
The competing event (optional). A list of numbers (typically integers) that represent a unit of time (typically days).

withdraw_events
The withdraw event (optional). A list of numbers (typically integers) that represent a unit of time (typically days).

followup_time
The follow-up time period (optinal). A number (typically and integer). 'end of study' will be the assigned censor event if any event (main, competing or withdraw events) has the value equal the followup_time.

main_label
(optional) Label for the main event (default = "main event").

competing_label
(optional) Label for the competing risk event (default = "competing event").

withdraw_label
(optional) Label for the withdraw event (default = "withdraw form study")

units
(optional) Time unit (e.g. "days", default = NA) #' @return data frame with two variables: a) 'censor_event' - factor with levels and labels of the different potential first events: 1 = "main event" 2 = "competing event" 3 = "withdraw from study" 4 = "end of study" b) 'censor_time' - time to event. Number (usually an integer). Returned data.frame length = length of original lists. Event date = followup_tme (censored), if no event.

Examples

# NOT RUN {
#' # load the test data in the bllFlow package
  df <- as.data.frame(read.csv("inst/extdata/RESPECT-EOL_validation.csv"))

# or
#  df <- RESPECT-EOL_validation

# if you don't have the bllFlow package loded.
 source(file.path(getwd(), 'R/make-event-dates.R'))

# create event data. Each new variable is a list of events that is the same
# length, nrow(df), as the orginial data, df.
  death    <- make_event_date(nrow(df), df$risk, 1825, label = 'death', units = 'days')
  competing   <- make_event_date(nrow(df), 0.2, 1825, label = 'censor', units = 'days')
  withdraw <- make_event_date(nrow(df), 0.40, 1825, label = 'withdraw', units = 'days')

# now find with event occurred first and assign the follow-up time to that event.
  censor <- assign_first_event(death, competing_events = competing,
    withdraw_events = withdraw, followup_time = 1825, units = units(death))
# the return is a data.frame (censor) with two new variables:
# censor$censor_event and censor$censor_time
summary(censor)

# you can add (mutate) the new variables to the original data
library(dplyr)
df <- df %>% bind_cols(censor)

# }