Calibration Plot Creates a calibration plot using plotly package. Returns an actual vs predicted plot. Used for model fit visualization

Calibration Plot Creates a calibration plot using plotly package. Returns an actual vs predicted plot. Used for model fit visualization

calibration_plot(data, observed, predicted, error=0, title=NA)

Arguments

data
a data.frame

observed
variable of observed observation from data.frame, object of class string

predicted
variable of predicted observations from data.frame, object of class string

error
percent error for error bars (optional)

title
an overall title for the plot (optional), object of class string

Examples

#upload MPorT table one data set
mport <- read.csv("https://raw.githubusercontent.com/Big-Life-Lab/bllFlow/data-visualization/inst/extdata/MPoRT-table-one.csv?token=ApmTWVvRI4hGeBRlPLR6R3P6J73Po5Paks5cWzfswA%3D%3D",
fileEncoding="UTF-8-BOM")

#> Warning: cannot open URL 'https://raw.githubusercontent.com/Big-Life-Lab/bllFlow/data-visualization/inst/extdata/MPoRT-table-one.csv?token=ApmTWVvRI4hGeBRlPLR6R3P6J73Po5Paks5cWzfswA%3D%3D': HTTP status was '404 Not Found'

#> Error in file(file, "rt", encoding = fileEncoding): cannot open the connection to 'https://raw.githubusercontent.com/Big-Life-Lab/bllFlow/data-visualization/inst/extdata/MPoRT-table-one.csv?token=ApmTWVvRI4hGeBRlPLR6R3P6J73Po5Paks5cWzfswA%3D%3D'

#create character vector of columns needed for visualizations
col <- c("observed_risk_1_year", "predicted_risk_1_year", "group_by_sex", "group_by_1", "group_by_2", "group_by_value_label_1")

#create new data frame using filter_table_1 helper function 
data<-filter_table_1(mport, col, groupby_sex="Female", groupby_1="age")

#> Error in filter_table_1(mport, col, groupby_sex = "Female", groupby_1 = "age"): could not find function "filter_table_1"

calibration_plot(data, "observed_risk_1_year", "predicted_risk_1_year", title="Observed vs Predicted for Female age")

#> Error in calibration_plot(data, "observed_risk_1_year", "predicted_risk_1_year",     title = "Observed vs Predicted for Female age"): could not find function "calibration_plot"