Cleans the data using the min and outlier columns in the variables sheet of the MSW. Outlier method is applied on a row if any of the variable values for that row is less than the min value as specified in the variables sheet. Outlier checking for the column is not applied if min value is NA.

Currently supported outlier methods are: Delete - Specified as 'delete' in MSW. Deletes the row from the data. number of deleted rows as well as their reason for deletion is stored in the metaData variable under the deletedRows name. Missing - Specified as 'missing' in MSW. Column value for that row which does not meet the criteria is set to NA. Not Applicable - TODO. Set to value - Specified as a number value in MSW. Column value for the row is set to the value specified in the outlier column.

# S3 method for BLLFlow
clean.Min(bllFlowModel, print = FALSE, ...)

Arguments

bllFlowModel
bllFlow model created using the BLLFlow method

print
If set true, prints the following metrics about the model: 1. Number of rows deleted from the dataset

...
Arguments for next methods in the chain

Value

bllFlowModel that has had its data modified by the paramaters located in the variables object

Examples

# Install the packages

# Read in the data we will use

library(survival)
data(pbc)

# Read in the MSW and variable_details sheet for the PBC model
variablesSheet <- read.csv(file.path(getwd(),
 'bllFlow/extdata/PBC/PBC-variables.csv'))

#> Warning: cannot open file '/Users/yulrics/Code/Project Big Life/bllFlow/docs/reference/bllFlow/extdata/PBC/PBC-variables.csv': No such file or directory

#> Error in file(file, "rt"): cannot open the connection

variableDetailsSheet <- read.csv(file.path(getwd(),
'bllFlow/extdata/PBC/PBC-variable-details.csv'))

#> Warning: cannot open file '/Users/yulrics/Code/Project Big Life/bllFlow/docs/reference/bllFlow/extdata/PBC/PBC-variable-details.csv': No such file or directory

#> Error in file(file, "rt"): cannot open the connection

# Create a bllFlow R object for the PBC model using the above variables as args
library(bllFlow)

#> Error in library(bllFlow): there is no package called ‘bllFlow’

pbcModel <- BLLFlow(pbc, variablesSheet, variableDetailsSheet)

#> Error in BLLFlow(pbc, variablesSheet, variableDetailsSheet): object 'variablesSheet' not found

# passing non dataframe objects when generating the bllFlow object will cause errors
#pbcModel <- BLLFlow(pbc, c(1,2,3), variableDetailsSheet)

# Clean the data
cleanedPbcModel <- clean.Min(pbcModel)

#> Error in clean.Min(pbcModel): object 'pbcModel' not found

# if u wish to be updated in the log on what the function does set print to true
cleanedPbcModel <- clean.Min(cleanedPbcModel, print=TRUE)

#> Error in clean.Min(cleanedPbcModel, print = TRUE): object 'cleanedPbcModel' not found