Add an event date, when you only have an event probability

To create example data for algorithm development.

Use case: You have validation scoring data with event probablity but not event date or censoring date.

Use case: You have cohort data with only exposure variables and you would like to create a fictious event or censoring events.

make_event_dates(num_events, event_probability, followup_time,
  label = "time to event", units = NA)

Arguments

num_events
(required) Number of events (rows) to be caluculated. Should always equal the number of events in your orginal data. See example.

event_probability
(required) Target event probablity (number value from 0 to 1 representing the event probability.e.g. 5-year risk of death is 0.2 represents 20 % proabality of dying over 5 years or 1825 days.

followup_time
(required) Time period for event probablity (number, usually an integer).

label
(optional) Name of event. Default = 'Time to event'.

units
(optional) Time unit for event. e.g. "days" Default = NA #' @return List with event times (integers) Returned list length = list length of df. Event date = totalTime (censored), if no event.

Examples

# load the test data in the bllFlow package

# NOT RUN {
#loaod and example data.frame of observations which includes a calculation of
 risk estimate (risk for death at end-of-life).
 df <- as.data.frame(read.csv('inst/extdata/RESPECT-EOL_validation.csv'))
 # or 
 # df <- RESPECT-EOL_validation
 df_length <- nrow(df)
 
# each ovservation as event probablity of df$risk
  df$tt_event <- make_event_date(df_length, df$risk, 1825, units = 'days')
  
 # add variable labeles
   df$tt_event <- make_event_date(df_length, df$risk, 1825, 
   label = "end-of-life", units = 'days')
  
# check the label
  label(df$tt_event)
 
# check to see if events have been added
  table(df$tt_event)
 
# check to see if df$tt_event is about the same as df$risk
  mean(df$risk)
  nrow(df[df$tt_event < 1825,]) / nrow(df)

# make other event data like censoring or competing events
  df$ttcensor <- make_event_date(df_length, .01, 1825)
# }